adjusted wrist pos data size in robotic arm

+++ code/RoboticArm/lib/WristControl/WristControl.cpp
@@ -4,6 +4,6 @@ void wristControlSetup(){
     return;
 }
 
-void controlWrist(uint8_t* wrist_pos){
+void controlWrist(uint32_t* wrist_pos){
     return;
 }
